refactor : `MenuGroup` 패키지 생성 및 파일 이동
